tests/smoke/requirements.txt
PyYAML~=6.0.1